% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/associations.R
\name{chisq_pvalues}
\alias{chisq_pvalues}
\title{Compute Chi-squared p-values}
\usage{
chisq_pvalues(
  m_data,
  response,
  adjust_method = "fdr",
  mlog10_transform = TRUE,
  n_cores = 1,
  ...
)
}
\arguments{
\item{m_data}{Data matrix of observations by variables}

\item{response}{Response vector of length the number of observations}

\item{adjust_method}{Multiple testing p-value adjustment method.
Passed to stats::p.adjust. 'fdr' by default.}

\item{mlog10_transform}{Logical, transform p-values by minus log10.
True by default.}

\item{n_cores}{Number of cores}

\item{...}{Passed to stats::chisq.test}
}
\value{
Chi-squared p-values
}
\description{
Compute Chi-squared p-values
}
