% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/associations.R
\name{chisq_pvalues_gdata}
\alias{chisq_pvalues_gdata}
\title{Compute Chi-squared p-values on a Genotype data object}
\usage{
chisq_pvalues_gdata(
  gdata,
  snp_idxs,
  response_column = "region",
  response_value = "Europe",
  threshold = 2,
  ...
)
}
\arguments{
\item{gdata}{Genotype data object}

\item{snp_idxs}{SNPs indexes}

\item{response_column}{Response column in gdata scans annotations data frame}

\item{response_value}{Response value. The response vector will be a logical,
true if equal to the value, false otherwise.}

\item{threshold}{Keep only associations greater than the threshold}

\item{...}{Passed to chisq_pvalues}
}
\value{
SNPs annotation data frame, chi-squared p-values in column pvalues
}
\description{
Compute Chi-squared p-values on a Genotype data object
}
