\name{crohn}
\docType{data}
\alias{crohn}
\title{Crohn's disease data}

\description{
The data set consist of 103 common (>5\% minor allele frequency) SNPs 
genotyped in 129 trios from an European-derived population. These 
SNPs are in a 500-kb region on human chromosome 5q31 implicated as 
containing a genetic risk factor for Crohn disease.

Imported from the gap R package.

An example use of the data is with the following paper,
Kelly M. Burkett, Celia M. T. Greenwood, BradMcNeney, Jinko Graham.
Gene genealogies for genetic association mapping, with application to Crohn's disease.
Fron Genet 2013, 4(260) doi: 10.3389/fgene.2013.00260
}
\usage{data(crohn)}

\format{A data frame containing 387 rows and 212 columns}

\source{
MJ Daly, JD Rioux, SF Schaffner, TJ Hudson, ES Lander (2001) 
High-resolution haplotype structure in the human genome Nature 
Genetics 29:229-232
}
  
\keyword{datasets}
