% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplots_ld.R
\name{diamond_annots}
\alias{diamond_annots}
\title{Get diamond ggplot layer.}
\usage{
diamond_annots(data, x = "x", y = "y", color = "color", size = 0.5)
}
\arguments{
\item{data}{Data frame of 3 columns defining the diamonds}

\item{x}{Name of the column for horizontal positions}

\item{y}{Name of the column for vertical positions}

\item{color}{Name of the column for color values}

\item{size}{Radius of the diamonds}
}
\value{
gglayers
}
\description{
Diamond ggplot layer for ggplot_ld
}
