% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_methods.R
\name{fetch_allele1.GdsGenotypeReader}
\alias{fetch_allele1.GdsGenotypeReader}
\title{Fetch allele 1 (GdsGenotypeReader object)}
\usage{
\method{fetch_allele1}{GdsGenotypeReader}(obj, snps_idx)
}
\arguments{
\item{obj}{GenotypeData object}

\item{snps_idx}{SNPs indexes}
}
\value{
Allele 1
}
\description{
Fetch allele 1 (GdsGenotypeReader object)
}
