% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_methods.R
\name{fetch_allele1.GenotypeData}
\alias{fetch_allele1.GenotypeData}
\title{Fetch allele 1 (GenotypeData object)}
\usage{
\method{fetch_allele1}{GenotypeData}(obj, ...)
}
\arguments{
\item{obj}{GenotypeData object}

\item{...}{Passed to getAlleleA}
}
\value{
Allele 1
}
\description{
Fetch allele 1 (GenotypeData object)
}
