% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5_methods_genotype_data_subset.R
\name{fetch_allele1.GenotypeDataSubset}
\alias{fetch_allele1.GenotypeDataSubset}
\title{Fetch allele 1 (GenotypeDataSubset object)}
\usage{
\method{fetch_allele1}{GenotypeDataSubset}(obj, snps_idx)
}
\arguments{
\item{obj}{GenotypeDataSubset object}

\item{snps_idx}{SNPs indexes}
}
\value{
Allele 1
}
\description{
Fetch allele 1 (GenotypeDataSubset object)
}
