% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_methods.R
\name{fetch_allele2.default}
\alias{fetch_allele2.default}
\title{Fetch allele 2 (default object)}
\usage{
\method{fetch_allele2}{default}(obj, snps_idx)
}
\arguments{
\item{obj}{Default object}

\item{snps_idx}{SNPs indexes}
}
\description{
Fetch allele 2 (default object)
}
