% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_methods.R
\name{fetch_allele2.GdsGenotypeReader}
\alias{fetch_allele2.GdsGenotypeReader}
\title{Fetch allele 2 (GdsGenotypeReader object)}
\usage{
\method{fetch_allele2}{GdsGenotypeReader}(obj, snps_idx)
}
\arguments{
\item{obj}{GenotypeData object}

\item{snps_idx}{SNPs indexes}
}
\value{
Allele 2
}
\description{
Fetch allele 2 (GdsGenotypeReader object)
}
