% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_methods.R
\name{fetch_allele2.GenotypeData}
\alias{fetch_allele2.GenotypeData}
\title{Fetch allele 2 (GenotypeData object)}
\usage{
\method{fetch_allele2}{GenotypeData}(obj, ...)
}
\arguments{
\item{obj}{GenotypeData object}

\item{...}{Passed to getAlleleB}
}
\value{
Allele 2
}
\description{
Fetch allele 2 (GenotypeData object)
}
