% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_methods.R
\name{fetch_gds.default}
\alias{fetch_gds.default}
\title{Fetch GDS (default)}
\usage{
\method{fetch_gds}{default}(obj, ...)
}
\arguments{
\item{obj}{Default object}

\item{...}{Not passed}
}
\description{
Fetch GDS (default)
}
