% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_methods.R
\name{fetch_gds.GdsGenotypeReader}
\alias{fetch_gds.GdsGenotypeReader}
\title{Fetch GDS (GdsGenotypeReader)}
\usage{
\method{fetch_gds}{GdsGenotypeReader}(obj, ...)
}
\arguments{
\item{obj}{GdsGenotypeReader object}

\item{...}{Not passed}
}
\value{
S4 slot 'handler' of obj
}
\description{
Fetch GDS (GdsGenotypeReader)
}
