% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_methods.R
\name{fetch_gds.GenotypeData}
\alias{fetch_gds.GenotypeData}
\title{Fetch GDS (GenotypeData)}
\usage{
\method{fetch_gds}{GenotypeData}(obj, ...)
}
\arguments{
\item{obj}{GenotypeData object}

\item{...}{Not passed}
}
\value{
fetch_gds output on S4 slot 'data' of obj
}
\description{
Fetch GDS (GenotypeData)
}
