% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_methods.R
\name{fetch_gds.GenotypeDataSubset}
\alias{fetch_gds.GenotypeDataSubset}
\title{Fetch GDS (GenotypeDataSubset)}
\usage{
\method{fetch_gds}{GenotypeDataSubset}(obj, ...)
}
\arguments{
\item{obj}{GenotypeDataSubset object}

\item{...}{Not passed}
}
\description{
Fetch GDS (GenotypeDataSubset)
}
