% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomart_annots.R
\name{gdata_add_gene_annots}
\alias{gdata_add_gene_annots}
\title{gdata_add_gene_annots}
\usage{
gdata_add_gene_annots(
  gdata,
  snp_idxs,
  rsids_colname = "probe_id",
  biomart_metadb = get_biomart_metadb()
)
}
\arguments{
\item{gdata}{Genotype Data object}

\item{snp_idxs}{SNP indexes}

\item{rsids_colname}{Column of SNP annotation data frame with rs identifiers}

\item{biomart_metadb}{List with slots snpmart and ensembl, corresponding to 
the biomart databases to query for SNP identifiers
and gene names, respectively. See get_biomart_metadb
function.}
}
\value{
Genotype Data object
}
\description{
Add biomaRt gene annotations to Genotype Data object.
}
