% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomart_annots.R
\name{gdata_add_gene_annots_aim_example}
\alias{gdata_add_gene_annots_aim_example}
\title{gdata_add_gene_annots_aim_example}
\usage{
gdata_add_gene_annots_aim_example(gdata, aim_idxs)
}
\arguments{
\item{gdata}{Genotype Data object}

\item{aim_idxs}{AIM indexes in the example Genotype data object}
}
\value{
Genotype Data object
}
\description{
Add ancestry informative markers gene annotations to Genotype Data object.
Convenience function for the vignette to avoid querying biomaRt on build.
}
