% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomart_annots.R
\name{gdata_add_gene_annots_hladr_example}
\alias{gdata_add_gene_annots_hladr_example}
\title{gdata_add_gene_annots_hladr_example}
\usage{
gdata_add_gene_annots_hladr_example(gdata, hla_dr_idxs)
}
\arguments{
\item{gdata}{Genotype Data object}

\item{hla_dr_idxs}{HLA-DR indexes in the example Genotype data object}
}
\value{
Genotype Data object
}
\description{
Add HLA-DR gene annotations to Genotype Data object.
Convenience function for the vignette to avoid querying biomaRt on build.
}
