% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geno_data_fetch.R
\name{gdata_scans_annots}
\alias{gdata_scans_annots}
\title{gdata_scan_annots}
\usage{
gdata_scans_annots(gdata, scan_ids)
}
\arguments{
\item{gdata}{Genotype Data object}

\item{scan_ids}{Scan identifiers to subset}
}
\value{
Scans annotations data frame
}
\description{
Get scans annotations from a Genotype Data object or a subset.
}
