% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geno_data_fetch.R
\name{gdata_snps_annots}
\alias{gdata_snps_annots}
\title{gdata_snp_annots}
\usage{
gdata_snps_annots(gdata, snp_ids = NULL)
}
\arguments{
\item{gdata}{Genotype Data object}

\item{snp_ids}{SNP identifiers to subset}
}
\value{
SNP annotation data frame
}
\description{
Get SNPs annotations from a Genotype Data object or a subset.
}
