% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_methods.R
\name{get_scan_annot.GenotypeData}
\alias{get_scan_annot.GenotypeData}
\title{Get scans annotations (GenotypeData object)}
\usage{
\method{get_scan_annot}{GenotypeData}(obj, ...)
}
\arguments{
\item{obj}{GenotypeData object}

\item{...}{Not passed}
}
\value{
Data frame
}
\description{
Get scans annotations (GenotypeData object)
}
