% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_methods.R
\name{get_scan_annot.GenotypeDataSubset}
\alias{get_scan_annot.GenotypeDataSubset}
\title{Get scans annotations (GenotypeDataSubset object)}
\usage{
\method{get_scan_annot}{GenotypeDataSubset}(obj, ...)
}
\arguments{
\item{obj}{GenotypeDataSubset object}

\item{...}{Not passed}
}
\value{
Data frame
}
\description{
Get scans annotations (GenotypeDataSubset object)
}
