% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_methods.R
\name{get_snp_annot.GenotypeDataSubset}
\alias{get_snp_annot.GenotypeDataSubset}
\title{Get SNPs annotations (GenotypeDataSubset object)}
\usage{
\method{get_snp_annot}{GenotypeDataSubset}(obj, ...)
}
\arguments{
\item{obj}{GenotypeDataSubset object}

\item{...}{Not passed}
}
\value{
Data frame
}
\description{
Get SNPs annotations (GenotypeDataSubset object)
}
