% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtable_ld.R
\name{ggplot_snp_pos}
\alias{ggplot_snp_pos}
\title{Ggplot SNPs position}
\usage{
ggplot_snp_pos(
  df_snp,
  upper_subset = NULL,
  labels_colname = NULL,
  colors = snp_position_colors(nrow(df_snp))
)
}
\arguments{
\item{df_snp}{SNP annotation data frame with a column named position
and, if specified, one named as the labels_colname
parameter.}

\item{upper_subset}{Subset of df_snp for the positions on the upper side}

\item{labels_colname}{Optional column name of df_snp to use as SNP labels.}

\item{colors}{Colors for each SNP}
}
\value{
ggplot
}
\description{
Get SNPs position ggplot with mappings to combine with other ggplots.
Optionally add labels and an upper subset.
}
