% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_methods.R
\name{is_snp_first_dim.default}
\alias{is_snp_first_dim.default}
\title{Is SNP first dimension (default)}
\usage{
\method{is_snp_first_dim}{default}(obj, ...)
}
\arguments{
\item{obj}{Default object}

\item{...}{Not passed}
}
\value{
NA
}
\description{
Is SNP first dimension (default)
}
