% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_methods.R
\name{is_snp_first_dim.gds.class}
\alias{is_snp_first_dim.gds.class}
\title{Is SNP first dimension (GDS object)}
\usage{
\method{is_snp_first_dim}{gds.class}(obj, ...)
}
\arguments{
\item{obj}{GDS object}

\item{...}{Not passed}
}
\value{
Logical, TRUE if SNP is first dimension
}
\description{
Is SNP first dimension (GDS object)
}
