% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_methods.R
\name{is_snp_first_dim.GdsGenotypeReader}
\alias{is_snp_first_dim.GdsGenotypeReader}
\title{Is SNP first dimension (GdsGenotypeReader object)}
\usage{
\method{is_snp_first_dim}{GdsGenotypeReader}(obj, ...)
}
\arguments{
\item{obj}{GdsGenotypeReader object}

\item{...}{Not passed}
}
\value{
is_snp_first_dim output on S4 slot 'handler'
}
\description{
Is SNP first dimension (GdsGenotypeReader object)
}
