% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_methods.R
\name{is_snp_first_dim.GenotypeData}
\alias{is_snp_first_dim.GenotypeData}
\title{Is SNP first dimension (GenotypeData object)}
\usage{
\method{is_snp_first_dim}{GenotypeData}(obj, ...)
}
\arguments{
\item{obj}{Genotype data object}

\item{...}{Not passed}
}
\value{
is_snp_first_dim output on S4 slot 'data'
}
\description{
Is SNP first dimension (GenotypeData object)
}
