% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_methods.R
\name{is_snp_first_dim.MatrixGenotypeReader}
\alias{is_snp_first_dim.MatrixGenotypeReader}
\title{Is SNP first dimension (MatrixGenotypeReader object)}
\usage{
\method{is_snp_first_dim}{MatrixGenotypeReader}(obj, ...)
}
\arguments{
\item{obj}{MatrixGenotypeReader object}

\item{...}{Not passed}
}
\value{
TRUE
}
\description{
Is SNP first dimension (MatrixGenotypeReader object)
}
