% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_methods.R
\name{is_snp_first_dim.NcdfGenotypeReader}
\alias{is_snp_first_dim.NcdfGenotypeReader}
\title{Is SNP first dimension (NcdfGenotypeReader object)}
\usage{
\method{is_snp_first_dim}{NcdfGenotypeReader}(obj, ...)
}
\arguments{
\item{obj}{NcdfGenotypeReader object}

\item{...}{Not passed}
}
\value{
TRUE
}
\description{
Is SNP first dimension (NcdfGenotypeReader object)
}
