% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geno_data_fetch.R
\name{load_gds_as_genotype_data}
\alias{load_gds_as_genotype_data}
\title{Load GDS as Genotype Data}
\usage{
load_gds_as_genotype_data(
  gds_file,
  read_snp_annot = TRUE,
  read_scan_annot = TRUE
)
}
\arguments{
\item{gds_file}{Path of snpgds file}

\item{read_snp_annot}{Read the SNPs' annotations}

\item{read_scan_annot}{Read the scans' annotations}
}
\value{
Genotype Data object
}
\description{
Open a connection to a snpgds file (cf. SNPRelate package) as a Genotype
Data object.
}
