% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/associations.R
\name{parallel_apply}
\alias{parallel_apply}
\title{Separate a matrix in a list of matrices of length the number of cores and
apply a function on the columns in parallel}
\usage{
parallel_apply(m_data, apply_fun, n_cores = 1, ...)
}
\arguments{
\item{m_data}{Data matrix}

\item{apply_fun}{Function to apply}

\item{n_cores}{Number of cores}

\item{...}{Passed to apply_fun}
}
\value{
apply_fun return
}
\description{
Separate a matrix in a list of matrices of length the number of cores and
apply a function on the columns in parallel
}
