% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print_qc_as_tex_table}
\alias{print_qc_as_tex_table}
\title{print_qc_as_tex_table}
\usage{
print_qc_as_tex_table(
  gdata_qc,
  label = "qc",
  caption = paste("Quality control and feature selection of the subset of the",
    "human genome diversity project dataset.")
)
}
\arguments{
\item{gdata_qc}{Genotype Data object object returned by snprelate_qc}

\item{label}{Label of the Tex table}

\item{caption}{Caption of the Tex table}
}
\value{
Prints knitr::kable object using cat
}
\description{
Print information about quality control performed by the snprelate_qc
function.
}
