% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_hgdp.R
\name{save_hgdp_as_gds}
\alias{save_hgdp_as_gds}
\title{save_hgdp_as_gds}
\usage{
save_hgdp_as_gds(paths = hgdp_filepaths(), outpath = tempfile(), ...)
}
\arguments{
\item{paths}{Paths of the zip, txt, and gds files}

\item{outpath}{Output GDS file path}

\item{...}{Passed to save_genotype_data_as_gds}
}
\value{
Path of the saved gds file
}
\description{
Save the HGDP SNP data text file as a Genomic Data Structure file
}
