% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomart_annots.R
\name{select_region_idxs}
\alias{select_region_idxs}
\title{select_region_idxs}
\usage{
select_region_idxs(
  gdata,
  chromosome,
  position_min = -Inf,
  position_max = Inf,
  n_snps = 0,
  offset = 0
)
}
\arguments{
\item{gdata}{Genotype Data object}

\item{chromosome}{Chromosome to select}

\item{position_min}{Minimum base pair position to select}

\item{position_max}{Maximum base pair position to select}

\item{n_snps}{Maximum number of SNPs to return}

\item{offset}{Number of SNPs to offset}
}
\value{
SNP indexes of Genotype Data object
}
\description{
Select SNP indexes corresponding to a specific genomic region.
}
