% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snplinkage-package.R
\docType{package}
\name{snplinkage-package}
\alias{snplinkage}
\alias{snplinkage-package}
\title{snplinkage: Single Nucleotide Polymorphisms Linkage Disequilibrium Visualizations}
\description{
Linkage disequilibrium visualizations of up to several hundreds of single nucleotide polymorphisms (SNPs), annotated with chromosomic positions and gene names. Two types of plots are available for small numbers of SNPs (<40) and for large numbers (tested up to 500). Both can be extended by combining other ggplots, e.g. association studies results, and functions enable to directly visualize the effect of SNP selection methods, as minor allele frequency filtering and TagSNP selection, with a second correlation heatmap. The SNPs correlations are computed on Genotype Data objects from the 'GWASTools' package using the 'SNPRelate' package, and the plots are customizable 'ggplot2' and 'gtable' objects and are annotated using the 'biomaRt' package. Usage is detailed in the vignette with example data and results from up to 500 SNPs of 1,200 scans are in Charlon T. (2019) \doi{10.13097/archive-ouverte/unige:161795}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://gitlab.com/thomaschln/snplinkage}
  \item Report bugs at \url{https://gitlab.com/thomaschln/snplinkage/-/issues}
}

}
\author{
\strong{Maintainer}: Thomas Charlon \email{charlon@protonmail.com} (\href{https://orcid.org/0000-0001-7497-0470}{ORCID})

Authors:
\itemize{
  \item Karl Forner
  \item Alessandro Di Cara
  \item Jérôme Wojcik
}

}
\keyword{internal}
