% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snprelate_ld.R
\name{snprelate_ld}
\alias{snprelate_ld}
\title{Wrapper for snpgdsLDMat to compute r2}
\usage{
snprelate_ld(
  gdata,
  window_size = 0,
  min_r2 = 0,
  snps_idx = NULL,
  scans_idx = NULL,
  threads = 1,
  quiet = FALSE
)
}
\arguments{
\item{gdata}{A GenotypeData object}

\item{window_size}{Max number of SNPs in LD window, 0 for no window}

\item{min_r2}{Minimum r2 value to report}

\item{snps_idx}{Indices of snps to use}

\item{scans_idx}{Indices of scans to use}

\item{threads}{The number of threads to use}

\item{quiet}{Whether to be quiet}
}
\value{
A data frame with columns SNP_A, SNP_B, R2 for r2 >= min_r2
}
\description{
Wrapper for snpgdsLDMat to compute r2
}
